Renames remote data model
