+++ lib/src/database/models/remote_clinic_history.dart
